backend/requirements.txt
fastapi
sqlmodel
psycopg2-binary
uvicorn
python-dotenv
httpx
pytest
